make it idempotent

+++ internal/controller/prefectserver_controller.go
@@ -179,6 +179,7 @@ func (r *PrefectServerReconciler) reconcileMigrationJob(ctx context.Context, ser
 
 	foundMigrationJob := &batchv1.Job{}
 	err = r.Get(ctx, types.NamespacedName{Namespace: server.Namespace, Name: desiredMigrationJob.Name}, foundMigrationJob)
+
 	if errors.IsNotFound(err) {
 		log.Info("Creating migration Job", "name", desiredMigrationJob.Name)
 		if err = r.Create(ctx, desiredMigrationJob); err != nil {
@@ -202,22 +203,22 @@ func (r *PrefectServerReconciler) reconcileMigrationJob(ctx context.Context, ser
 
 		return &ctrl.Result{}, errors.NewBadRequest(errorMessage)
 	} else if migrationJobNeedsUpdate(&foundMigrationJob.Spec, &desiredMigrationJob.Spec, log) {
-		log.Info("Recreating migration Job", "name", desiredMigrationJob.Name)
+		log.Info("migration Job needs to be updated and replaced", "name", desiredMigrationJob.Name)
 		condition = conditions.NeedsUpdate(objName)
 
-		// deletes the job + recreates it.
+		// deletes the Job, so it can be recreated on the following reconciliation.
 		// k8s Jobs are immutable and need to be recreated if they are to be re-run.
-		if err = r.Delete(ctx, foundMigrationJob); err != nil {
-			condition = conditions.NotDeleted(objName, err)
-			return &ctrl.Result{}, err
+		if foundMigrationJob.GetDeletionTimestamp() == nil {
+			log.Info("Deleting existing migration Job", "name", foundMigrationJob.Name)
+			if err = r.Delete(ctx, foundMigrationJob); err != nil {
+				condition = conditions.NotDeleted(objName, err)
+				return &ctrl.Result{}, err
+			}
+			condition = conditions.Deleted(objName)
+		} else {
+			log.Info("Waiting for migration Job to be deleted, so it can be recreated", "name", desiredMigrationJob.Name)
 		}
-
-		if err = r.Create(ctx, desiredMigrationJob); err != nil {
-			condition = conditions.NotCreated(objName, err)
-			return &ctrl.Result{}, err
-		}
-
-		condition = conditions.Created(objName)
+		return &ctrl.Result{Requeue: true}, nil
 	} else {
 		if !meta.IsStatusConditionTrue(server.Status.Conditions, "MigrationJobReconciled") {
 			condition = conditions.Updated(objName)
